Audit: six findings, and playlists you can take away again
Three were mine from this morning. Stale playlist tabs kept stealing
clicks after VJ mode was turned on, because both places that reset them
sat inside the branch that does not run in VJ mode. A MIDI file's
all-notes-off went to "first available" rather than the port it had been
playing to, which is the one case where the cue is already gone and
cannot be asked. And the parsed-file cache handed out pointers into a
vector that reallocates.

Three were already there. Undo's debounce compared deck 0 and nothing
else, so with more than one playlist an edit on any other deck was never
recorded and could not be undone. VJ mode claimed every output, so
enabling it silently discarded every routing decision in the show; it
owns the programme output now. And a layer cap had borrowed the deck
limit, which was a safe number by coincidence.

Playlists can be removed as well as added, each with its own sound. The
remap is written down once and everything that holds a deck index goes
through it -- outputs' hosts and stacks, master assignments, target
cues, VJ's A and B -- because a reference that is not remapped does not
break, it starts meaning the neighbour. Default names follow their new
positions; a name somebody typed is left alone.

The routing menu now says it in the order James described it: an output,
then a layer in that output.

+++ native/app/app_quick_action.ipp
@@ -103,7 +103,7 @@
       case QuickAction::DashSlotDelete: {
         const int at = param;
         if (at < 0 || at >= static_cast<int>(project_.dashboard.size())) break;
-        const Uint32 now = SDL_GetTicks();
+        const Uint64 now = SDL_GetTicks();
         if (dashDeleteArmedSlot_ != at || now - dashDeleteArmedAtMs_ > 4000) {
           dashDeleteArmedSlot_ = at;
           dashDeleteArmedAtMs_ = now;
